+++ impl/rb/lib/uxid.rb
@@ -14,6 +14,12 @@ module UXID
   CROCKFORD_ENCODING = "0123456789ABCDEFGHJKMNPQRSTVWXYZ"
   DELIMITER = "_"
 
+  SIZE_WARNING_MESSAGE = <<~EOM
+    WARN: The UXID :size param is for Symbol T-Shirt sizes only.
+          Please use the :size param with Symbols (:xs, :s, :m, :l, :xl, or :xsmall, :small, :medium, :large, :xlarge),
+          or use the :rand_size param for specific size of random bytes.
+  EOM
+
   class Error < StandardError; end
 
   def self.generate attrs = {}
@@ -25,8 +31,38 @@ module UXID
     model = ::UXID::Model.new
 
     model.time = attrs[:time] || Time.now
-    model.size = attrs[:size] || 10
     model.prefix = attrs[:prefix] || ""
+
+    if attrs[:rand_size]
+      model.rand_size = attrs[:rand_size]
+    else
+      model.size = attrs[:size]
+      case model.size
+      when Integer
+        puts SIZE_WARNING_MESSAGE
+        model.rand_size = model.size
+
+      when String
+        puts SIZE_WARNING_MESSAGE
+        model.rand_size = 10
+
+      when :xs, :xsmall
+        model.rand_size = 0
+
+      when :s, :small
+        model.rand_size = 2
+
+      when :m, :medium
+        model.rand_size = 5
+
+      when :l, :large
+        model.rand_size = 7
+
+      else
+        model.rand_size = 10
+      end
+
+    end
 
     model
   end